checked and updated and tested the api

+++ src/main/java/com/foodservice/controller/DeliveryDriverController.java
@@ -111,7 +111,7 @@ public class DeliveryDriverController {
      }
      
      @GetMapping("/{orderId}/driver")
-     public ResponseEntity<ApiResponseDTO> getDriverByOrder(@PathVariable Long orderId) {
+     public ResponseEntity<ApiResponseDTO> getDriverByOrder(@PathVariable Integer orderId){
 
          DeliveryDriverResponseDTO driver = orderService.getDriverByOrder(orderId);
